data_ingestion_realtime/requirements.txt
kafka-python==2.0.2
happybase==1.2.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
